moving tests to component
