+++ cypress/component/basic/MyPlugin.js
