+++ cypress/component/button/ButtonCounter.vue
