+++ cypress/component/counter-vuex/Counter.vue
