+++ cypress/component/counter-vuex/store.js
